Create function to check validity of indices files.
Note, this is not currently used when GITM is running within the SWMF. In standalone GITM, indices files are checked to ensure data coverage over the entire run. But within the SWMF, GITM cannot see the start/end times of the simulation when reading the Param file. This will be fixed eventually...

Creating this subroutine to allow the code to compile & sync to GITM's util/ directory.